backend/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
jinja2==3.1.3
sqlalchemy==2.0.40
websockets>=13.0,<15.0
bcrypt==4.1.2
python-dotenv==1.0.1
aiofiles==23.2.1
google-genai
